outputs '_' by the number of times specified by the user

+++ 0x04-more_functions_nested_loops/6-print_line.c
@@ -7,7 +7,7 @@
  */
 void print_line(int n)
 {
-	unsigned short int x;
+	int x;
 
 	for (x = 0; x < n; x++)
 	{
@@ -23,3 +23,4 @@ void print_line(int n)
 	_putchar(10);
 	return (0);
 }
+
